refactor: move constants to utils, upd imports, remove commented code
