+++ my-app/pages/index.js
@@ -6,10 +6,9 @@ import { Games } from '../components/Games';
 import { Controls } from '../components/Controls/Controls';
 import { useRouter } from 'next/router';
 import Head from 'next/head';
-import { API_URL, KEY_URL } from '../utils/constants';
+import { API_URL, initialFilters, KEY_URL } from '../utils/constants';
 import { Loader } from '../components/Loader';
 import { useRouteLoading } from '../hooks/useRouteLoading';
-import { initialFilters } from '../constants/constants';
 
 export default function Home({ initial }) {
   const router = useRouter();
